225: Accept multibyte character as wide character literal
On most Unix-like systems, wide character literal is 32-bit long
and encodes a Unicode code point. On Windows, that is 16-bit
long and encodes a UTF-16 code unit. Clearly, there's a portability
issue here.

Being said that, some header files contain wide character literal,
so we need to support that so that toycc can include such files.

We assume that source files are always encoded in UTF-8.